+++ lib/src/validators/number.dart
@@ -1,13 +1,13 @@
 import 'package:acanthis/src/operations/checks.dart';
 
 /// Number check for less than or equal to a value.
-class LteNumberCheck extends AcanthisCheck<num> {
-  final num value;
+class LteNumberCheck<T extends num> extends AcanthisCheck<T> {
+  final T value;
 
   LteNumberCheck(
     this.value, {
     String? message,
-    String Function(num value)? messageBuilder,
+    String Function(T value)? messageBuilder,
   }) : super(
          error:
              messageBuilder?.call(value) ??
@@ -17,19 +17,19 @@ class LteNumberCheck extends AcanthisCheck<num> {
        );
 
   @override
-  bool call(num value) {
+  bool call(T value) {
     return value <= this.value;
   }
 }
 
 /// Number check for greater than or equal to a value.
-class GteNumberCheck extends AcanthisCheck<num> {
-  final num value;
+class GteNumberCheck<T extends num> extends AcanthisCheck<T> {
+  final T value;
 
   GteNumberCheck(
     this.value, {
     String? message,
-    String Function(num value)? messageBuilder,
+    String Function(T value)? messageBuilder,
   }) : super(
          error:
              messageBuilder?.call(value) ??
@@ -39,21 +39,21 @@ class GteNumberCheck extends AcanthisCheck<num> {
        );
 
   @override
-  bool call(num value) {
+  bool call(T value) {
     return value >= this.value;
   }
 }
 
 /// Number check for between two values.
-class BetweenNumberCheck extends AcanthisCheck<num> {
-  final num min;
-  final num max;
+class BetweenNumberCheck<T extends num> extends AcanthisCheck<T> {
+  final T min;
+  final T max;
 
   BetweenNumberCheck(
     this.min,
     this.max, {
     String? message,
-    String Function(num min, num max)? messageBuilder,
+    String Function(T min, T max)? messageBuilder,
   }) : super(
          error:
              messageBuilder?.call(min, max) ??
@@ -63,19 +63,19 @@ class BetweenNumberCheck extends AcanthisCheck<num> {
        );
 
   @override
-  bool call(num value) {
+  bool call(T value) {
     return value >= min && value <= max;
   }
 }
 
 /// Number check for Greater than a value.
-class GtNumberCheck extends AcanthisCheck<num> {
-  final num value;
+class GtNumberCheck<T extends num> extends AcanthisCheck<T> {
+  final T value;
 
   GtNumberCheck(
     this.value, {
     String? message,
-    String Function(num value)? messageBuilder,
+    String Function(T value)? messageBuilder,
   }) : super(
          error:
              messageBuilder?.call(value) ??
@@ -85,19 +85,19 @@ class GtNumberCheck extends AcanthisCheck<num> {
        );
 
   @override
-  bool call(num value) {
+  bool call(T value) {
     return value > this.value;
   }
 }
 
 /// Number check for Less than a value.
-class LtNumberCheck extends AcanthisCheck<num> {
-  final num value;
+class LtNumberCheck<T extends num> extends AcanthisCheck<T> {
+  final T value;
 
   LtNumberCheck(
     this.value, {
     String? message,
-    String Function(num value)? messageBuilder,
+    String Function(T value)? messageBuilder,
   }) : super(
          error:
              messageBuilder?.call(value) ??
@@ -107,57 +107,57 @@ class LtNumberCheck extends AcanthisCheck<num> {
        );
 
   @override
-  bool call(num value) {
+  bool call(T value) {
     return value < this.value;
   }
 }
 
 /// Number check for positive value.
-class PositiveNumberCheck extends AcanthisCheck<num> {
+class PositiveNumberCheck<T extends num> extends AcanthisCheck<T> {
   const PositiveNumberCheck({String? message})
     : super(error: message ?? 'Value must be positive', name: 'positive');
 
   @override
-  bool call(num value) {
+  bool call(T value) {
     return value > 0;
   }
 }
 
 /// Number check for negative value.
-class NegativeNumberCheck extends AcanthisCheck<num> {
+class NegativeNumberCheck<T extends num> extends AcanthisCheck<T> {
   const NegativeNumberCheck({String? message})
     : super(error: message ?? 'Value must be negative', name: 'negative');
 
   @override
-  bool call(num value) {
+  bool call(T value) {
     return value < 0;
   }
 }
 
 /// Number check for nonpositive value.
-class NonPositiveNumberCheck extends AcanthisCheck<num> {
+class NonPositiveNumberCheck<T extends num> extends AcanthisCheck<T> {
   const NonPositiveNumberCheck({String? message})
     : super(error: message ?? 'Value must be nonpositive', name: 'nonpositive');
 
   @override
-  bool call(num value) {
+  bool call(T value) {
     return value <= 0;
   }
 }
 
 /// Number check for nonnegative value.
-class NonNegativeNumberCheck extends AcanthisCheck<num> {
+class NonNegativeNumberCheck<T extends num> extends AcanthisCheck<T> {
   const NonNegativeNumberCheck({String? message})
     : super(error: message ?? 'Value must be nonnegative', name: 'nonnegative');
 
   @override
-  bool call(num value) {
+  bool call(T value) {
     return value >= 0;
   }
 }
 
 /// Number check for multiple of a value.
-class MultipleOfCheck extends AcanthisCheck<num> {
+class MultipleOfCheck<T extends num> extends AcanthisCheck<T> {
   final int value;
 
   MultipleOfCheck(
@@ -173,85 +173,85 @@ class MultipleOfCheck extends AcanthisCheck<num> {
        );
 
   @override
-  bool call(num value) {
+  bool call(T value) {
     return value % this.value == 0;
   }
 }
 
 /// Number check for integer value.
-class IntegerNumberCheck extends AcanthisCheck<num> {
+class IntegerNumberCheck<T extends num> extends AcanthisCheck<T> {
   const IntegerNumberCheck({String? message})
     : super(error: message ?? 'Value must be an integer', name: 'integer');
 
   @override
-  bool call(num value) {
+  bool call(T value) {
     return value is int;
   }
 }
 
 /// Number check for double value.
-class DoubleNumberCheck extends AcanthisCheck<num> {
+class DoubleNumberCheck<T extends num> extends AcanthisCheck<T> {
   const DoubleNumberCheck({String? message})
     : super(error: message ?? 'Value must be a double', name: 'double');
 
   @override
-  bool call(num value) {
+  bool call(T value) {
     return value is! int;
   }
 }
 
 /// Number check for finite value.
-class FiniteNumberCheck extends AcanthisCheck<num> {
+class FiniteNumberCheck<T extends num> extends AcanthisCheck<T> {
   const FiniteNumberCheck({String? message})
     : super(error: message ?? 'Value must be finite', name: 'finite');
 
   @override
-  bool call(num value) {
+  bool call(T value) {
     return value.isFinite;
   }
 }
 
 /// Number check for infinite value.
-class InfiniteNumberCheck extends AcanthisCheck<num> {
+class InfiniteNumberCheck<T extends num> extends AcanthisCheck<T> {
   const InfiniteNumberCheck({String? message})
     : super(error: message ?? 'Value must be infinite', name: 'infinite');
 
   @override
-  bool call(num value) {
+  bool call(T value) {
     return value.isInfinite;
   }
 }
 
 /// Number check for NaN value.
-class NaNNumberCheck extends AcanthisCheck<num> {
+class NaNNumberCheck<T extends num> extends AcanthisCheck<T> {
   const NaNNumberCheck({String? message})
     : super(error: message ?? 'Value must be NaN', name: 'nan');
 
   @override
-  bool call(num value) {
+  bool call(T value) {
     return value.isNaN;
   }
 }
 
 /// Number check for not NaN value.
-class NotNaNNumberCheck extends AcanthisCheck<num> {
+class NotNaNNumberCheck<T extends num> extends AcanthisCheck<T> {
   const NotNaNNumberCheck({String? message})
     : super(error: message ?? 'Value must not be NaN', name: 'notNaN');
 
   @override
-  bool call(num value) {
+  bool call(T value) {
     return !value.isNaN;
   }
 }
 
 /// Number check for enumerated values.
-class EnumeratedNumberCheck extends AcanthisCheck<num> {
-  final List<num> values;
+class EnumeratedNumberCheck<T extends num> extends AcanthisCheck<T> {
+  final List<T> values;
 
   EnumeratedNumberCheck(
     this.values, {
     String? message,
-    String Function(List<num> values)? messageBuilder,
+    String Function(List<T> values)? messageBuilder,
   }) : super(
          error:
              messageBuilder?.call(values) ??
@@ -261,7 +261,7 @@ class EnumeratedNumberCheck extends AcanthisCheck<num> {
        );
 
   @override
-  bool call(num value) {
+  bool call(T value) {
     return values.contains(value);
   }
 }